Updated parameter type for get_with_key_vec
Before it took `&[&str]`, which meant if I wanted to pass in, for
example, a `Vec<String>`, it was a whole ordeal. This is first solved
with the trait `AsRef<str>`. However, because of where I'm using it, I
also needed to add `serde_yaml::mapping::Index` as a trait bound.

The conversion before was something like:

    vec.iter().map(|s| s.as_str()).collect:<Vec<&str>>().as_slice()

And now it's none of that.