Updated keypath implementation
I realized that there was no way to get the components. In doing that, I
decided to instead store a `String`. Calling `components()` yields a
Vector over slices into that string. Also, before storing the string,
each component is run through `trim()`.

I spent a bunch of time looking at messing around with iterators and
such, and in this particular case this feels like a good way to do it. I
just parse and store the string once. I generate a `Vec<&str>` on-demand
with `components()`, which can be used, presumably as long as `KeyPath`
exists, though it's interesting it doesn't need a lifetime specifier.