Converted to a double-ended iterator
I wanted to be able to use `rev()` on the returned iterator. It should
have been trivial, but it wasn't because I was using an inclusive range.
That's a weird gotcha. In any case, I hacked around it with a +1 but
then I had to hide a lint relevant to it.